+++ src/app/components/collection-text-types/facsimiles/facsimiles.component.ts
@@ -1,4 +1,4 @@
-import { ChangeDetectionStrategy, Component, DestroyRef, Injector, effect, inject, input, output, signal } from '@angular/core';
+import { Component, DestroyRef, Injector, effect, inject, input, output, signal } from '@angular/core';
 import { NgStyle } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { AlertButton, AlertController, AlertInput, IonButton, IonFab, IonFabButton, IonIcon, IonInput, IonSpinner, ModalController } from '@ionic/angular';
@@ -25,7 +25,6 @@ import { sortArrayOfObjectsNumerically } from '@utility-functions';
   templateUrl: './facsimiles.component.html',
   styleUrls: ['./facsimiles.component.scss'],
   imports: [NgStyle, FormsModule, IonButton, IonFab, IonFabButton, IonIcon, IonInput, IonSpinner, DraggableImageDirective, TrustHtmlPipe],
-  changeDetection: ChangeDetectionStrategy.OnPush,
   host: { ngSkipHydration: 'true' }
 })
 export class FacsimilesComponent {