+++ src/app/components/collection-text-types/illustrations/illustrations.component.ts
@@ -1,4 +1,4 @@
-import { ChangeDetectionStrategy, Component, DestroyRef, Injector, NgZone, afterRenderEffect, computed, effect, inject, input, output, signal } from '@angular/core';
+import { Component, DestroyRef, Injector, NgZone, afterRenderEffect, computed, effect, inject, input, output, signal } from '@angular/core';
 import { takeUntilDestroyed, toObservable } from '@angular/core/rxjs-interop';
 import { NgClass } from '@angular/common';
 import { IonButton, IonSpinner, ModalController } from '@ionic/angular';
@@ -20,8 +20,7 @@ import { ScrollService } from '@services/scroll.service';
   selector: 'illustrations',
   templateUrl: './illustrations.component.html',
   styleUrls: ['./illustrations.component.scss'],
-  imports: [NgClass, IonButton, IonSpinner],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  imports: [NgClass, IonButton, IonSpinner]
 })
 export class IllustrationsComponent {
   // ─────────────────────────────────────────────────────────────────────────────