+++ src/app/components/collection-text-types/legend/legend.component.ts
@@ -1,4 +1,4 @@
-import { ChangeDetectionStrategy, Component, DestroyRef, ElementRef, Injector, LOCALE_ID, NgZone, Renderer2, afterRenderEffect, computed, inject, input, signal, untracked } from '@angular/core';
+import { Component, DestroyRef, ElementRef, Injector, LOCALE_ID, NgZone, Renderer2, afterRenderEffect, computed, inject, input, signal, untracked } from '@angular/core';
 import { takeUntilDestroyed, toObservable } from '@angular/core/rxjs-interop';
 import { IonSpinner } from '@ionic/angular';
 import { catchError, Observable, of, switchMap, tap } from 'rxjs';
@@ -16,8 +16,7 @@ import { ScrollService } from '@services/scroll.service';
   selector: 'text-legend',
   templateUrl: './legend.component.html',
   styleUrls: ['./legend.component.scss'],
-  imports: [IonSpinner, TrustHtmlPipe],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  imports: [IonSpinner, TrustHtmlPipe]
 })
 export class LegendComponent {
   // ─────────────────────────────────────────────────────────────────────────────