+++ src/app/components/collection-text-types/manuscripts/manuscripts.component.ts
@@ -1,4 +1,4 @@
-import { ChangeDetectionStrategy, Component, DestroyRef, ElementRef, Injector, afterRenderEffect, computed, effect, inject, input, output, signal, untracked } from '@angular/core';
+import { Component, DestroyRef, ElementRef, Injector, afterRenderEffect, computed, effect, inject, input, output, signal, untracked } from '@angular/core';
 import { takeUntilDestroyed, toObservable } from '@angular/core/rxjs-interop';
 import { AlertButton, AlertController, AlertInput, IonButton, IonIcon, IonSpinner, IonToggle } from '@ionic/angular';
 import { catchError, of, switchMap, tap } from 'rxjs';
@@ -20,8 +20,7 @@ import { ViewOptionsService } from '@services/view-options.service';
   selector: 'manuscripts',
   templateUrl: './manuscripts.component.html',
   styleUrls: ['./manuscripts.component.scss'],
-  imports: [IonButton, IonIcon, IonSpinner, IonToggle, TrustHtmlPipe],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  imports: [IonButton, IonIcon, IonSpinner, IonToggle, TrustHtmlPipe]
 })
 export class ManuscriptsComponent {
   // ─────────────────────────────────────────────────────────────────────────────