+++ src/app/components/collection-text-types/reading-text/reading-text.component.ts
@@ -1,4 +1,4 @@
-import { ChangeDetectionStrategy, Component, DestroyRef, ElementRef, Injector, NgZone, Renderer2, afterRenderEffect, computed, inject, input, output, signal, untracked } from '@angular/core';
+import { Component, DestroyRef, ElementRef, Injector, NgZone, Renderer2, afterRenderEffect, computed, inject, input, output, signal, untracked } from '@angular/core';
 import { takeUntilDestroyed, toObservable } from '@angular/core/rxjs-interop';
 import { IonSpinner, ModalController } from '@ionic/angular';
 import { catchError, combineLatest, map, of, switchMap, tap } from 'rxjs';
@@ -23,8 +23,7 @@ import { enableFrontMatterPageOrTextViewType, isFileNotFoundHtml } from '@utilit
   selector: 'reading-text',
   templateUrl: './reading-text.component.html',
   styleUrls: ['./reading-text.component.scss'],
-  imports: [IonSpinner, MathJaxDirective, TrustHtmlPipe],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  imports: [IonSpinner, MathJaxDirective, TrustHtmlPipe]
 })
 export class ReadingTextComponent {
   // ─────────────────────────────────────────────────────────────────────────────