+++ src/app/components/collection-text-types/variants/variants.component.ts
@@ -1,4 +1,4 @@
-import { ChangeDetectionStrategy, Component, DestroyRef, ElementRef, Injector, afterRenderEffect, computed, effect, inject, input, output, signal, untracked } from '@angular/core';
+import { Component, DestroyRef, ElementRef, Injector, afterRenderEffect, computed, effect, inject, input, output, signal, untracked } from '@angular/core';
 import { takeUntilDestroyed, toObservable } from '@angular/core/rxjs-interop';
 import { AlertButton, AlertController, AlertInput, IonButton, IonIcon, IonSpinner } from '@ionic/angular';
 import { catchError, of, switchMap, tap } from 'rxjs';
@@ -20,8 +20,7 @@ import { ViewOptionsService } from '@services/view-options.service';
   selector: 'variants',
   templateUrl: './variants.component.html',
   styleUrls: ['./variants.component.scss'],
-  imports: [IonButton, IonIcon, IonSpinner, TrustHtmlPipe],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  imports: [IonButton, IonIcon, IonSpinner, TrustHtmlPipe]
 })
 export class VariantsComponent {
   // ─────────────────────────────────────────────────────────────────────────────